rename: store 폴더명 변경
store폴더 명에 앞이 빈칸이 하나 들어가서 이를 수정함
