+++ haul-fe/src/App.jsx
@@ -4,7 +4,7 @@ import { ThemeProvider } from "styled-components";
 import theme from "./styles/theme/Theme.jsx";
 import "./assets/fonts/fonts.css";
 import ToastRoot from "./components/Toast/ToastRoot.jsx";
-import { ReservationStoreProvider } from "./\bstore/reservationStore.jsx";
+import { ReservationStoreProvider } from "./store/reservationStore.jsx";
 
 function App() {
   return (
